shell.CommandHandler.handle(): display own output using writer
Each CommandHandler implementation is responsible for writing its
own output to the PrintWriter rather than returning it to the
CommandPrompt caller. This allows messages to be displayed more
immediately and/or progressively should a command take a
perceptible amount of time to complete.

CommandHandlers that need to write output should Inject the PrintWriter
Bean as writer.